Core: AddonReader: Rename resetEvent to DataReady - This is an event Producer.
Core: GoapAgent: In short, the GoapAgent thread tightly syncronized by the AddonThread.

Core: GoapAgent: rename manualReset to sessionPauseEvent. On the other hand removed the fixed Thread.Sleep(1) which likely to be the culprit why was driven to madness after hours of debugging in the past 2 years. Since the GoapAgent was polling 4x more compare to the AddonThread, the GoapAgent might be ended up in a state when after a wait.Update() call the expected cause and effect was not happened. This issue most likely manifested as various issues like targeting an NPC which supposed to be blacklisted - or certain actions which had limited time window was never executed.